requirements.txt
matplotlib==3.7.3
opencv_python==4.8.1.78
Pillow==9.3.0
torch==2.1.0+cpu
torchvision==0.16.0+cpu
